% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{compare_two_way_ranks}
\alias{compare_two_way_ranks}
\title{Compare all ranks in a list taken two at a time}
\usage{
compare_two_way_ranks(ranks)
}
\arguments{
\item{ranks}{A named list with a vector in each slot, with a specific order.}
}
\value{
A data.frame with the following columns:
- \code{steps}: the window size of the calculated rank;
- \code{step_fraction}: \code{size} but divided by the length of the vectors;
Each additional column is a comparison, with the name
\verb{<one>_vs_<two>}, where \verb{<one>} and \verb{<two>} are the names of the
two considered vectors in the \code{ranks} list.
}
\description{
All ranks must be of the same length. Their names in the list are the names
used in the output, with the pattern \code{one_vs_two}.
}
\author{
Hedmad
}
