% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{test_comparisons}
\alias{test_comparisons}
\title{Compute a p-value of the difference in ordering between two vectors a and b}
\usage{
test_comparisons(a, b, perms = 1000)
}
\arguments{
\item{a}{A vector to compare.}

\item{b}{Another vector to compare with.}

\item{perms}{The number of permutations to compute. Defaults to 1000.}
}
\value{
A p-value between 0 and 1.
}
\description{
The test involves permuting \code{b} a number \code{perms} times, and then computing
the score for each comparison.
The final p-value is the fraction of permutation scores that are greater
than the true score divided by \code{perms}.
}
\details{
This is incredibly expensive, so it will take a long, long time to compute
for long vectors and many permutations.

The \code{a} and \code{b} vectors must have the same length.
}
\author{
Hedmad
}
