% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{window_compare_rank}
\alias{window_compare_rank}
\title{Compare the order of the items in two vectors}
\usage{
window_compare_rank(one, two, binsize = 1)
}
\arguments{
\item{one}{First vector to compare}

\item{two}{Second vector to compare}

\item{binsize}{The size of the rolling window. Defaults to 1}
}
\value{
a value between 0 and 1, where 0 is perfect difference and 1 is
perfect equality.
}
\description{
Run a rolling window from the start of both vectors and compute how many
items in each window of \code{one} are in the same window of \code{two}, and so on.
For each individual "hit", increment a counter by one.
Then, divide by the total hits by the total length of all windows.
}
\details{
This function is symmetrical: if you swap \code{one} and \code{two}, the result is the
same.

It is designed to compare two lists that have the same elements but in
different order. However, it is not a requirement.

\code{length(one) == length(two)} MUST be true.
}
\examples{
window_compare_rank(c("a", "b", "c", "d"), c("d", "a", "c", "b"), binsize = 2)

}
\author{
Hedmad
}
